Check that upserted data contain the key information
When doing an upsert, if the user does not indicate the keys, it
defaults to ('ID', 'value'). If these keys are not part of the upserted
rows, this will update all the table with the column values of the last
passed rows.

This PR checks that the keys are present in the passed rows.